Moved afterCommit handling to dedicated file
