+++ src/packages/emmett/src/eventStore/afterCommit/index.ts
@@ -1,0 +1,1 @@
+export * from './afterEventStoreCommitHandler';
