+++ src/packages/emmett/src/eventStore/index.ts
@@ -1,5 +1,6 @@
-export * from './eventStore';
+export * from './afterCommit';
 export * from './events';
+export * from './eventStore';
 export * from './expectedVersion';
 export * from './inMemoryEventStore';
 export * from './subscriptions';
